refactor: move entry point logic from main to server
